refactor: separate LogInForm component, add logon request result if it fails
