+++ src/pages/login.tsx
@@ -1,49 +1,15 @@
 import { Layout } from "~/component/layout/layout";
 import Link from "next/link";
-import { useRouter } from "next/router";
-import Cookie from "js-cookie";
 import { type GetServerSideProps } from "next";
-import { AuthForm } from "~/component/auth/authForm";
-import { AUTHORIZATION_TOKEN_KEY } from "~/utils/auth/authorizationTokenKey";
 import { AuthLayout } from "~/component/auth/authLayout";
 import { withoutAuth } from "~/utils/auth/withoutAuth";
+import { LogInForm } from "~/component/auth/logInForm";
 
-const Login = () => {
-  const router = useRouter();
-
-  async function verifyLogin(login: string, password: string) {
-    try {
-      const response = await fetch("/api/auth/login", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ login, password }),
-      });
-
-      type AuthorizationResponse = {
-        message: string;
-        token: string;
-      };
-
-      if (response.ok) {
-        const data = (await response.json()) as AuthorizationResponse;
-        Cookie.set(AUTHORIZATION_TOKEN_KEY, data.token, {
-          expires: 30,
-          path: "/",
-        });
-        await router.push("/jobs");
-      } else {
-        const errorData = (await response.json()) as AuthorizationResponse;
-        console.error(errorData.message);
-      }
-    } catch (error) {
-      console.error("Network error:", error);
-    }
-  }
-
+export default function Login() {
   return (
     <Layout>
       <AuthLayout authorizationType="Log in">
-        <AuthForm handleFormSubmit={verifyLogin} authorizationType="Log in" />
+        <LogInForm />
         <p className="mt-10 text-center text-sm text-gray-500">
           Not a member?{" "}
           <Link
@@ -56,8 +22,6 @@ const Login = () => {
       </AuthLayout>
     </Layout>
   );
-};
+}
 
 export const getServerSideProps: GetServerSideProps = withoutAuth();
-
-export default Login;
